global_chem/global_chem_extensions/requirements.txt
global-chem
rdkit-pypi
pandas
requests
numpy



